Added new endpoint for reports.

+++ src/main/java/br/com/livraria/repository/LivroRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.livraria.model.Livro;
 
-public interface LivroRepository extends JpaRepository<Livro, Integer> {
+public interface LivroRepository extends JpaRepository<Livro, Long> {
 
 	
 }
